Create Calculator (addition, subtraction, multiplication and division)
Implement the simplest calculator, supporting four operations: +, -, *, /. We only need to consider the case where the input and output are integers, and the data and operation results will not exceed the range represented by int. Input two integers and an operator (+, -, *, /) and output the operation result.
However: (1) If the divisor is 0, the output is: Divided by zero!
(2) If an invalid operator occurs (that is, it is not one of +, -, *, /), the output is: Invalid operator!