Create What year is it today?.cpp
Input a positive integer m, representing the year. If m is a leap year, output yes, otherwise output no. The judgment rule for defining a leap year is: Most years that are evenly divisible by 4 are leap years, but years that are evenly divisible by 100 but not 400 are not leap years, and years that are evenly divisible by 3200 are not leap years. For example, 1900 is not a leap year, 2000 is a leap year, and 3200 is not a leap year.