Create Three Days Later.cpp
What day will it be in three days?

Input: Three positive integers a, b, and c indicate that today is day b, month c, year a, which ensures that the day exists legally.

Output: The three numbers x y z indicate that three days after the input time is x year y month z day. Separate each number with a space, and do not break the line at the end.