Create Buy House
Describe: A programmer starts working with an annual salary of RMB 10,000. He hopes to buy a house in Xiamen. The current price is 2 million. Assume that the house price increases at a rate of B% per year, and the programmer's annual salary remains unchanged in the future, and he does not eat or drink. , no need to pay taxes, and all the income of RMB 10,000 is saved every year. I want to ask if I can buy this house within 20 years? If so, in what year can I buy this house? (Annual salary of RMB 10,000 in the first year and house price of RMB 2 million)

Input: A positive integer n represents the number of cases.
Each set of cases consists of two positive integers a and b (10<=a<=50, 1<=b<=20)

Output: If the house can be bought in the 20th year or before, then an integer M is output, indicating that it needs to be bought in the Mth year at the earliest, otherwise No is output.
Each set of cases must be line-wrapped after output.