Create Star Array - Barbell
Input: There is only one set of cases, consisting of 2 positive integers a and b, where b is an odd number and a>=2b+1. (a<=30, b<=11)

Output: Outputs a barbell shape consisting of * based on the values ​​of a and b. Do not output unnecessary spaces. There should be a newline after the last asterisk.

example: 10 3
***
***
***
 *
 *
 *
 *
***
***
***