Create Function 6.cpp
Define a function f with two int type parameters a and b, which is used to output asterisks in row a and column b. Input two integers x and y into the main function, and then call the f function to output the asterisks in the x row and y column.